+++ packages/terra-abstract-modal/CHANGELOG.md
@@ -3,6 +3,13 @@ Changelog
 
 Unreleased
 ----------
+### Changed
+* Update react/react-dom peer-dependency to latest stable minor release
+* Updated `rootProp` default value from `[data-terra-base]` to `#root`
+
+### Removed
+* Removed peer dependency on terra-base
+* Removed baseStyles import from terra-base
 
 2.4.0 - (March 12, 2019)
 ------------------