+++ packages/terra-application-utility/CHANGELOG.md
@@ -3,6 +3,15 @@ Changelog
 
 Unreleased
 ----------
+### Changed
+* Update react/react-dom peer-dependency to latest stable minor release
+
+### Changed
+* Background color for header utility example changed to increase contrast ratio
+
+### Removed
+* Removed peer dependency on terra-base
+* Removed baseStyles import from terra-base
 
 2.5.0 - (March 20, 2019)
 ------------------